azure_monitor/python_logger_opencensus_azure/monitoring/requirements.txt
opencensus-ext-azure
opencensus-ext-logging
opencensus-ext-flask
opencensus.ext.requests
flask
